+++ datafiles/tests/test_formats.py
@@ -12,22 +12,9 @@ def describe_serialize():
     def data():
         return {'key': "value", 'items': [1, 'a', None]}
 
-    def describe_ruamel_yaml():
+    def describe_yaml():
         def it_indents_blocks_by_default(expect, data):
             text = formats.serialize(data, '.yaml')
-            expect(text) == dedent(
-                """
-            key: value
-            items:
-              - 1
-              - a
-              -
-            """
-            )
-
-    def describe_pyyaml():
-        def it_indents_blocks_by_default(expect, data):
-            text = formats.serialize(data, '.yaml', formatter=formats.PyYAML)
             expect(text) == dedent(
                 """
             key: value
@@ -46,14 +33,9 @@ def describe_deserialize():
         path.write_text("")
         return path
 
-    def describe_ruamel_yaml():
+    def describe_yaml():
         def with_empty_file(expect, path):
             data = formats.deserialize(path, '.yaml')
-            expect(data) == {}
-
-    def describe_pyyaml():
-        def with_empty_file(expect, path):
-            data = formats.deserialize(path, '.yaml', formatter=formats.PyYAML)
             expect(data) == {}
 
     def describe_json():